+++ fms/distributed/tensorparallel.py
@@ -8,7 +8,7 @@ import torch.distributed._functional_collectives as funcol
 from torch import nn
 
 
-def apply_colwise_tp(par_mod: nn.Linear, mod: nn.Linear, world_size, rank):
+def apply_colwise_tp(par_mod: nn.Linear, mod: nn.Linear, rank: int, world_size: int):
     # Divide the weight matrix along the last dimension.
     output_size_per_partition = mod.out_features // world_size
     with torch.no_grad():
@@ -20,7 +20,7 @@ def apply_colwise_tp(par_mod: nn.Linear, mod: nn.Linear, world_size, rank):
     # print(f"For rank {rank}, we have the following weights: Base weight {mod.weight} bias {mod.bias}; Par weight {par_mod.weight}, bias {par_mod.bias}")
 
 
-def apply_rowwise_tp(par_mod: nn.Linear, mod: nn.Linear, world_size, rank):
+def apply_rowwise_tp(par_mod: nn.Linear, mod: nn.Linear, rank: int, world_size: int):
     # Divide the weight matrix along the last dimension.
     output_size_per_partition = mod.in_features // world_size
     with torch.no_grad():
@@ -35,7 +35,9 @@ def apply_rowwise_tp(par_mod: nn.Linear, mod: nn.Linear, world_size, rank):
     # print(f"For rank {rank}, we have the following weights: Base weight {mod.weight}, bias {mod.bias}; Par weight {par_mod.weight}, bias {par_mod.bias}")
 
 
-def apply_embedding_tp(par_mod: nn.Embedding, mod: nn.Embedding, world_size, rank):
+def apply_embedding_tp(
+    par_mod: nn.Embedding, mod: nn.Embedding, rank: int, world_size: int
+):
     # Divide the weight matrix along the last dimension.
     output_size_per_partition = mod.embedding_dim // world_size
     with torch.no_grad():
@@ -45,11 +47,12 @@ def apply_embedding_tp(par_mod: nn.Embedding, mod: nn.Embedding, world_size, ran
     # print(f"For rank {rank}, we have the following weights: Base weight {mod.weight} bias {mod.bias}; Par weight {par_mod.weight}, bias {par_mod.bias}")
 
 
-def apply_moe_tp(par_mod: nn.Module, mod: nn.Module, param_names, world_size, rank):
+def apply_moe_tp(
+    par_mod: nn.Module, mod: nn.Module, param_names, rank: int, world_size: int
+):
     with torch.no_grad():
         for param in param_names:
             par_param = getattr(par_mod, param)
-            print(param, par_param.shape[1] // world_size)
             par_param.copy_(
                 torch.split(
                     getattr(mod, param), par_param.shape[1] // world_size, dim=1
